updated offline features
