+++ package/offline/clip.py
@@ -1,0 +1,22 @@
+import os
+from package.offline import copy_to_clipboard
+from package.passsword import valid_password
+
+def clip(snippet_name, password):
+    if not isinstance(password, int) or not valid_password(password):
+        raise ValueError("Incorrect password")
+
+    base_dir = os.path.dirname(__file__)
+    snippets_dir = os.path.join(base_dir, "stash")
+    
+    try:
+        snippet_path = os.path.join(snippets_dir, snippet_name)
+
+        with open(snippet_path, "r") as file:
+            content = file.read()
+        copy_to_clipboard(content)
+        print("Snippet copied to clipboard.")
+    except FileNotFoundError:
+        print("Error: No file found with the specified name.")
+    except Exception as e:
+        print(f"Error: {e}")
