--- package/test.py
@@ -1,5 +1,0 @@
-from scrap.grab import write, show, clip
-
-write("fghd", "text.txt")
-show("fghd")
-clip("fghd")
